AP-01-003/004/005: record user starter implementations (003 and 005 buggy)
User's first-pass implementations for three chapter-1 problems, recorded
as-is per repo policy (buggy drafts committed with diagnosis).

AP-01-003 normalize_uint8:
  Changed `(images/255).astype(dtype)` to
  `(images/(255).astype(dtype)).astype(dtype,copy=True)` to perform the
  division in the target dtype before the final cast.
  BUG: `(255).astype(dtype)` raises AttributeError because 255 is a Python
  int, not a numpy array (int has no .astype method). Intent was likely
  `np.array(255).astype(dtype)` or `np.float32(255)`.
  Tests: 1/4 pass (only test_rejects_bad_dtypes); 3 ERROR.
  Benchmark: ERROR (same AttributeError).

AP-01-004 flatten_images:
  Implemented. require_independent=True path does
  `images.reshape(batchsize,-1).copy(order="C")` for an owned C-contiguous
  result. Shared path tries `reshape(batchsize,-1,copy=False)` and falls
  back to `.copy(order="C")` on ValueError, printing a diagnostic line.
  Tests: 3/3 pass.
  Benchmark: contiguous_shared_alias=True, contiguous_owned_alias=False,
  shared_mutation_reached_source=True, owned_bytes=12544.

AP-01-005 standardize_features:
  Implemented with explicit shape/dtype/finite validation and full-matrix
  construction: MeanMatrix via ones((N,1)) @ mean.reshape((1,D)), ScaleMatrix
  via np.diag of reciprocals (inverse-scale), result = (x - MeanMatrix) @ ScaleMatrix.
  BUG 1: `(mean).astype(np.float64)` on a Python float raises AttributeError
  (float has no .astype) - hits the scalar branch in test_scalar_statistics.
  BUG 2: `if not np.isfinite(mean):` on an ndarray raises "truth value of an
  array is ambiguous" - missing np.all() (used for x but forgotten for the
  mean/scale ndarray branches) - hits test_square_axis_ambiguity_uses_features.
  Tests: 1/3 pass (only test_rejects_accidental_shapes_and_bad_scale); 2 ERROR.
  Benchmark: ERROR (same ambiguous-truth-value bug).

Conclusion: AP-01-004 passes fully; AP-01-003 and AP-01-005 have runtime
bugs (int/float .astype calls, missing np.all on array finiteness checks)
to diagnose and fix next.

+++ chapters/chapter_01_data_contracts/AP-01-003_safe_pixel_normalization/starter.py
@@ -33,5 +33,5 @@ def normalize_uint8(
         raise ValueError(f"expected images of shape (batchsize,length,width) or (length,width), got {images.shape[-3:]}")
     if dtype not in (np.float64, np.float32):
         raise ValueError(f"expected dtype to be float32 or float64, got {dtype}")
-    result= (images/255).astype(dtype,copy=True)
+    result= (images/(255).astype(dtype)).astype(dtype,copy=True)
     return result
